# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  Income: 'Income',
  Expense: 'Expense',
  Investment: 'Investment',
  Saving: 'Saving',
} as const;

export interface Transaction {
  id: string;
  /** ISO date (YYYY-MM-DD) */
  date: string;
  description: string;
  category: string;
  type: TransactionType;
  amount: number;
  /** @nullable */
  notes: string | null;
  /** ISO timestamp */
  createdAt: string;
}

export interface TransactionInput {
  /** ISO date (YYYY-MM-DD) */
  date: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  type: TransactionType;
  /** @exclusiveMinimum 0 */
  amount: number;
  /** @nullable */
  notes?: string | null;
}

export interface BulkDeleteInput {
  ids: string[];
}

export interface Budget {
  category: string;
  limit: number;
  /** Amount spent this month for this category */
  spent: number;
}

export interface BudgetInput {
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  limit: number;
}

export interface MetricTotals {
  income: number;
  expenses: number;
  savings: number;
  investments: number;
}

export interface MonthlyPoint {
  /** Label like "Jan 2026" */
  month: string;
  income: number;
  expenses: number;
  savings: number;
  investments: number;
}

export interface CategoryBreakdownItem {
  category: string;
  amount: number;
}

export interface SavingsPoint {
  /** Label like "Jan 2026" */
  month: string;
  savings: number;
}

export interface Summary {
  currentMonth: MetricTotals;
  /** Last 6 months of income/expenses/savings/investments */
  monthlyTrend: MonthlyPoint[];
  /** Expense breakdown by category for the current month */
  expenseByCategory: CategoryBreakdownItem[];
  /** Savings over the last 12 months */
  savingsTrend: SavingsPoint[];
  /** Total invested across all time */
  totalInvested: number;
  /** Total invested this month */
  investedThisMonth: number;
  /** Total number of transactions (0 means empty/sample-data eligible) */
  transactionCount: number;
  /** True when the current data set is unmodified seeded sample data */
  isSampleData: boolean;
}

export type ListTransactionsParams = {
type?: TransactionType;
category?: string;
search?: string;
/**
 * Inclusive ISO date (YYYY-MM-DD) lower bound.
 */
startDate?: string;
/**
 * Inclusive ISO date (YYYY-MM-DD) upper bound.
 */
endDate?: string;
};