# lib/api-zod/src/generated/types/budget.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Budget {
  category: string;
  limit: number;
  /** Amount spent this month for this category */
  spent: number;
}