# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetInput {
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  limit: number;
}