# lib/api-zod/src/generated/types/savingsPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavingsPoint {
  /** Label like "Jan 2026" */
  month: string;
  savings: number;
}