# lib/api-zod/src/generated/types/summary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryBreakdownItem } from './categoryBreakdownItem';
import type { MetricTotals } from './metricTotals';
import type { MonthlyPoint } from './monthlyPoint';
import type { SavingsPoint } from './savingsPoint';

export interface Summary {
  currentMonth: MetricTotals;
  /** Last 6 months of income/expenses/savings/investments */
  monthlyTrend: MonthlyPoint[];
  /** Expense breakdown by category for the current month */
  expenseByCategory: CategoryBreakdownItem[];
  /** Savings over the last 12 months */
  savingsTrend: SavingsPoint[];
  /** Total invested across all time */
  totalInvested: number;
  /** Total invested this month */
  investedThisMonth: number;
  /** Total number of transactions (0 means empty/sample-data eligible) */
  transactionCount: number;
  /** True when the current data set is unmodified seeded sample data */
  isSampleData: boolean;
}