# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from './transactionType';

export interface Transaction {
  id: string;
  /** ISO date (YYYY-MM-DD) */
  date: string;
  description: string;
  category: string;
  type: TransactionType;
  amount: number;
  /** @nullable */
  notes: string | null;
  /** ISO timestamp */
  createdAt: string;
}