# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from './transactionType';

export interface TransactionInput {
  /** ISO date (YYYY-MM-DD) */
  date: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  type: TransactionType;
  /** @exclusiveMinimum 0 */
  amount: number;
  /** @nullable */
  notes?: string | null;
}