# lib/api-zod/src/generated/types/transactionType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  Income: 'Income',
  Expense: 'Expense',
  Investment: 'Investment',
  Saving: 'Saving',
} as const;